+++ mongodb-store/examples/mongodb.rs
@@ -4,9 +4,9 @@ use axum::{response::IntoResponse, routing::get, Router};
 use serde::{Deserialize, Serialize};
 use time::Duration;
 use tokio::{signal, task::AbortHandle};
-use tower_sessions::{Expiry, Session, SessionManagerLayer};
-use tower_sessions_core::ExpiredDeletion;
-use tower_sessions_mongodb_store::{mongodb::Client, MongoDBStore};
+use tower_sessions_ext::{Expiry, Session, SessionManagerLayer};
+use tower_sessions_ext_core::ExpiredDeletion;
+use tower_sessions_ext_mongodb_store::{mongodb::Client, MongoDBStore};
 
 const COUNTER_KEY: &str = "counter";
 
@@ -23,7 +23,7 @@ async fn handler(session: Session) -> impl IntoResponse {
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let database_url = std::option_env!("DATABASE_URL").expect("Missing DATABASE_URL.");
     let client = Client::with_uri_str(database_url).await?;
-    let session_store = MongoDBStore::new(client, "tower-sessions".to_string());
+    let session_store = MongoDBStore::new(client, "tower-sessions-ext".to_string());
 
     let deletion_task = tokio::task::spawn(
         session_store