+++ tests/test-integration.rs
@@ -4,8 +4,8 @@ mod common;
 #[cfg(test)]
 mod moka_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_moka_store::MokaStore;
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_moka_store::MokaStore;
 
     use crate::common::build_app;
 
@@ -21,8 +21,8 @@ mod moka_store_tests {
 #[cfg(test)]
 mod redis_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_redis_store::{fred::prelude::*, RedisStore};
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_redis_store::{fred::prelude::*, RedisStore};
 
     use crate::common::build_app;
 
@@ -47,8 +47,8 @@ mod redis_store_tests {
 #[cfg(test)]
 mod sqlite_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_sqlx_store::{sqlx::SqlitePool, SqliteStore};
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_sqlx_store::{sqlx::SqlitePool, SqliteStore};
 
     use crate::common::build_app;
 
@@ -67,8 +67,8 @@ mod sqlite_store_tests {
 #[cfg(test)]
 mod postgres_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_sqlx_store::{sqlx::PgPool, PostgresStore};
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_sqlx_store::{sqlx::PgPool, PostgresStore};
 
     use crate::common::build_app;
 
@@ -88,8 +88,8 @@ mod postgres_store_tests {
 #[cfg(test)]
 mod mysql_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_sqlx_store::{sqlx::MySqlPool, MySqlStore};
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_sqlx_store::{sqlx::MySqlPool, MySqlStore};
 
     use crate::common::build_app;
 
@@ -110,15 +110,15 @@ mod mysql_store_tests {
 #[cfg(test)]
 mod mongodb_store_tests {
     use axum::Router;
-    use tower_sessions::SessionManagerLayer;
-    use tower_sessions_mongodb_store::{mongodb, MongoDBStore};
+    use tower_sessions_ext::SessionManagerLayer;
+    use tower_sessions_ext_mongodb_store::{mongodb, MongoDBStore};
 
     use crate::common::build_app;
 
     async fn app(max_age: Option<Duration>) -> Router {
         let database_url = std::option_env!("MONGODB_URL").unwrap();
         let client = mongodb::Client::with_uri_str(database_url).await.unwrap();
-        let session_store = MongoDBStore::new(client, "tower-sessions".to_string());
+        let session_store = MongoDBStore::new(client, "tower-sessions-ext".to_string());
         let session_manager = SessionManagerLayer::new(session_store).with_secure(true);
 
         build_app(session_manager, max_age)
@@ -130,9 +130,9 @@ mod mongodb_store_tests {
 #[cfg(test)]
 mod caching_store_tests {
     use axum::Router;
-    use tower_sessions::{CachingSessionStore, SessionManagerLayer};
-    use tower_sessions_moka_store::MokaStore;
-    use tower_sessions_sqlx_store::{sqlx::SqlitePool, SqliteStore};
+    use tower_sessions_ext::{CachingSessionStore, SessionManagerLayer};
+    use tower_sessions_ext_moka_store::MokaStore;
+    use tower_sessions_ext_sqlx_store::{sqlx::SqlitePool, SqliteStore};
 
     use crate::common::build_app;