+++ mongodb-store/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 # 0.11.0
 
-- Update `tower-sessions` to `0.11.0`
+- Update `tower-sessions-ext` to `0.11.0`
 
 # 0.10.0
 
-- Unbundle from `tower-sessions`.
+- Unbundle from `tower-sessions-ext`.